Refactor application logic into discrete functions
This is of course a matter of taste, but when working on issue #2, I
first did this in order to better understand the flow of the program
and then reverted the changes in case you wouldn't like them. This
brings them back in a separate commit so you can separately decide if
you'd like to merge them.